+++ ProjectManagement/ProjectManagement.Client/Pages/Calculation/Form/ResourceFormUI.razor.cs
@@ -42,8 +42,8 @@ namespace ProjectManagement.Client.Pages.Calculation.Form
 
             //ResourceUpdate = r;
         }
-        async Task Update(double ch1, double ch2, double? BaseCost, double CapWaste, double? CO2, double Cost,
-            double? FixedQ, string Unit, int? AccountId, int? resTypeId, int? resTypeSortId)
+        async Task Update(decimal ch1, decimal ch2, decimal? BaseCost, decimal CapWaste, decimal? CO2, decimal Cost,
+            decimal? FixedQ, string Unit, int? AccountId, int? resTypeId, int? resTypeSortId)
         {
             ResourceUpdate.Data.ChangeFactor1 = ch1;
             ResourceUpdate.Data.ChangeFactor2 = ch2;
